+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [1.0.1]
 
 - Add support for new Agreements Engine parameters
- 
+
 ## [1.0.0]
 
 - Supports API changes from 20210324