+++ CHANGELOG.md
@@ -3,12 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.2]
 
-- Fix Agreements Engine parameter unscheduled_type
-- Add support for the new Apple Pay endpoints
+- Fix 'Agreements Engine' parameter 'unscheduled_type'
+- Add support for Apple Pay
 
 ## [1.0.1]
 
-- Add support for new Agreements Engine parameters
+- Add support for new 'Agreements Engine' parameters
 
 ## [1.0.0]